requirements.txt
aniso8601==10.0.1
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-RESTful==0.3.10
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
Flask-JWT-Extended==4.6.0
bcrypt==4.0.1
PyJWT==2.8.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-migrate==4.0.5
alembic==1.13.1

